add LocalWebpushrOptions
